+++ src/main/java/com/ubb/eventapp/repository/GroupRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Collection;
 
-public interface GroupRepository extends JpaRepository<Group, Integer> {
+public interface GroupRepository extends JpaRepository<Group, Long> {
 
     /**
      * Finds groups whose {@code nombre} contains the provided text, ignoring case.
@@ -25,7 +25,7 @@ public interface GroupRepository extends JpaRepository<Group, Integer> {
      * @return list of groups represented by the user
      */
     @Query("select gm.group from GroupMember gm where gm.user.id = :userId and gm.rolGrupo = com.ubb.eventapp.model.GroupRole.REPRESENTANTE")
-    List<Group> findByRepresentativeUser(@Param("userId") Integer userId);
+    List<Group> findByRepresentativeUser(@Param("userId") Long userId);
 
     /**
      * Finds groups that are associated with any of the provided tag ids.
