+++ src/main/java/com/ubb/eventapp/service/UserService.java
@@ -11,7 +11,7 @@ import java.util.Optional;
 public interface UserService {
     User registerUser(User user);
     User updateProfile(User user);
-    Optional<User> findById(Integer id);
+    Optional<User> findById(Long id);
 
     /**
      * Aggregates some statistics for the user's profile page.
@@ -19,7 +19,7 @@ public interface UserService {
      * @param userId id of the user to summarize
      * @return summary object with counts and trophies
      */
-    ProfileSummary getProfileSummary(Integer userId);
+    ProfileSummary getProfileSummary(Long userId);
 
     /**
      * Collects statistics about the user's event participation and creation.
@@ -27,7 +27,7 @@ public interface UserService {
      * @param userId id of the user
      * @return event summary information
      */
-    ProfileEvents getProfileEvents(Integer userId);
+    ProfileEvents getProfileEvents(Long userId);
 
     /**
      * Builds a calendar of all events the user has created, attended or plans
@@ -36,7 +36,7 @@ public interface UserService {
      * @param userId id of the user
      * @return list of calendar entries grouped by date
      */
-    java.util.List<CalendarEntry> getEventCalendar(Integer userId);
+    java.util.List<CalendarEntry> getEventCalendar(Long userId);
 
     /**
      * Lists the identifiers of future events the user plans to attend.
@@ -44,5 +44,5 @@ public interface UserService {
      * @param userId id of the user
      * @return wrapper object with event ids
      */
-    EventsToAttend getEventsToAttend(Integer userId);
+    EventsToAttend getEventsToAttend(Long userId);
 }
